fix(resume): move metadata to layout and fix client component
- Moved metadata export to layout.js
- Kept client-side code in page.js
- Fixed build error with 'use client' and metadata conflict
